+++ probables/__init__.py
@@ -1,6 +1,6 @@
 ''' pyprobables module '''
 from __future__ import (unicode_literals, absolute_import, print_function)
-from .blooms import (BloomFilter, BloomFilterOnDisk)
+from .blooms import (BloomFilter, BloomFilterOnDisk, CountingBloomFilter)
 from .countminsketch import (CountMinSketch, HeavyHitters, StreamThreshold,
                              CountMeanSketch, CountMeanMinSketch)
 from .exceptions import (InitializationError, NotSupportedError,
@@ -17,4 +17,4 @@ __bugtrack_url__ = 'https://github.com/barrust/pyprobables/issues'
 
 __all__ = ['BloomFilter', 'BloomFilterOnDisk', 'CountMinSketch',
            'HeavyHitters', 'StreamThreshold', 'CountMeanSketch',
-           'CountMeanMinSketch']
+           'CountMeanMinSketch', 'CountingBloomFilter']