+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # PyProbables Changelog
 
+### Version 0.0.6:
+* Probabilistic data structures added:
+    * Counting Bloom Filter
+* Minor code clean-up
+
 ### Version 0.0.5:
 * Better on-line documentation
 * Changed access to some public functions